clear node package
